Rename Old -> Classic, New -> Modern
