+++ src/components/tabs/antimatter-dimensions/ClassicAntimatterDimensionRow.vue
@@ -1,6 +1,11 @@
 <script>
+import PrimaryButton from "@/components/PrimaryButton";
+
 export default {
-  name: "NewAntimatterDimensionRow",
+  name: "ClassicAntimatterDimensionRow",
+  components: {
+    PrimaryButton
+  },
   props: {
     tier: {
       type: Number,
@@ -19,8 +24,7 @@ export default {
       singleCost: new Decimal(0),
       until10Cost: new Decimal(0),
       isAffordable: false,
-      buyUntil10: true,
-      howManyCanBuy: 0,
+      isAffordableUntil10: false,
       isContinuumActive: false,
       continuumValue: 0,
       isShown: false
@@ -29,9 +33,6 @@ export default {
   computed: {
     name() {
       return AntimatterDimension(this.tier).shortDisplayName;
-    },
-    costDisplay() {
-      return this.buyUntil10 ? format(this.until10Cost) : format(this.singleCost);
     },
     amountDisplay() {
       return this.tier < 8 ? format(this.amount, 2, 0) : formatInt(this.amount);
@@ -59,33 +60,34 @@ export default {
       if (tier > DimBoost.maxDimensionsUnlockable) return;
       const dimension = AntimatterDimension(tier);
       this.isUnlocked = dimension.isAvailableForPurchase;
-      const buyUntil10 = player.buyUntil10;
       this.isCapped = tier === 8 && Enslaved.isRunning && dimension.bought >= 1;
-      this.multiplier.copyFrom(AntimatterDimension(tier).multiplier);
+      this.multiplier.copyFrom(dimension.multiplier);
       this.amount.copyFrom(dimension.totalAmount);
+      this.totalAmount = dimension.totalAmount;
       this.bought = dimension.bought;
       this.boughtBefore10 = dimension.boughtBefore10;
-      this.howManyCanBuy = buyUntil10 ? dimension.howManyCanBuy : Math.min(dimension.howManyCanBuy, 1);
       this.singleCost.copyFrom(dimension.cost);
-      this.until10Cost.copyFrom(dimension.cost.times(Math.max(dimension.howManyCanBuy, 1)));
+      this.until10Cost.copyFrom(dimension.costUntil10);
       if (tier < 8) {
         this.rateOfChange.copyFrom(dimension.rateOfChange);
       }
       this.isAffordable = dimension.isAffordable;
-      this.buyUntil10 = buyUntil10;
+      this.isAffordableUntil10 = dimension.isAffordableUntil10;
       this.isContinuumActive = Laitela.continuumActive;
       if (this.isContinuumActive) this.continuumValue = dimension.continuumValue;
       this.isShown =
         (DimBoost.totalBoosts > 0 && DimBoost.totalBoosts + 3 >= tier) || PlayerProgress.infinityUnlocked();
     },
-    buy() {
+    buySingle() {
       if (this.isContinuumActive) return;
-      if (this.howManyCanBuy === 1) {
-        buyOneDimension(this.tier);
-      } else {
-        buyAsManyAsYouCanBuy(this.tier);
+      buyOneDimension(this.tier);
+      if (this.tier === 2) {
+        Tutorial.turnOffEffect(TUTORIAL_STATE.DIM2);
       }
-
+    },
+    buyUntil10() {
+      if (this.isContinuumActive) return;
+      buyManyDimension(this.tier);
       if (this.tier === 2) {
         Tutorial.turnOffEffect(TUTORIAL_STATE.DIM2);
       }
@@ -115,10 +117,10 @@ export default {
     :class="{ 'c-dim-row--not-reached': !isUnlocked }"
   >
     <div class="c-dim-row__label c-dim-row__name">
-      {{ name }} Antimatter D <span class="c-antimatter-dim-row__multiplier">{{ formatX(multiplier, 1, 1) }}</span>
+      {{ name }} Antimatter Dimension {{ formatX(multiplier, 1, 1) }}
     </div>
     <div class="c-dim-row__label c-dim-row__label--growable">
-      {{ amountDisplay }}
+      {{ amountDisplay }} ({{ formatInt(boughtBefore10) }})
       <span
         v-if="rateOfChange.neq(0)"
         class="c-dim-row__label--small"
@@ -126,43 +128,32 @@ export default {
         {{ rateOfChangeDisplay }}
       </span>
     </div>
-    <button
-      class="o-primary-btn o-primary-btn--new"
-      :class="{ 'o-primary-btn--disabled': (!isAffordable && !isContinuumActive) || !isUnlocked || isCapped}"
-      @click="buy"
+    <PrimaryButton
+      v-if="!isContinuumActive"
+      :enabled="isAffordable && !isCapped && isUnlocked"
+      class="o-primary-btn--buy-ad o-primary-btn--buy-single-ad l-dim-row__button"
+      :class="tutorialClass()"
+      :ach-tooltip="boughtTooltip"
+      @click="buySingle"
     >
-      <div
-        v-tooltip="boughtTooltip"
-        class="button-content"
-        :class="tutorialClass()"
-      >
-        <span v-if="isCapped">
-          Shattered by Enslaved
-        </span>
-        <span v-else-if="isContinuumActive">
-          Continuum:
-          <br>
-          {{ continuumString }}
-        </span>
-        <span v-else>
-          Buy {{ howManyCanBuy }}
-          <br>
-          Cost: {{ costDisplay }}
-        </span>
-      </div>
-      <div
-        v-if="!isContinuumActive && isUnlocked && !isCapped"
-        class="fill"
-      >
-        <div
-          class="fill-purchased"
-          :style="{ 'width': boughtBefore10*10 + '%' }"
-        />
-        <div
-          class="fill-possible"
-          :style="{ 'width': howManyCanBuy*10 + '%' }"
-        />
-      </div>
-    </button>
+      <span v-if="isCapped">Capped</span>
+      <template v-else>
+        <span v-if="showCostTitle(singleCost)">Cost: </span>{{ format(singleCost) }}
+      </template>
+    </PrimaryButton>
+    <PrimaryButton
+      :enabled="(isAffordableUntil10 || isContinuumActive) && !isCapped && isUnlocked"
+      class="o-primary-btn--buy-ad o-primary-btn--buy-10-ad l-dim-row__button"
+      :ach-tooltip="boughtTooltip"
+      @click="buyUntil10"
+    >
+      <span v-if="isCapped">Capped</span>
+      <span v-else-if="isContinuumActive">Continuum: {{ continuumString }}</span>
+      <template v-else>
+        Until {{ formatInt(10) }},
+        <span v-if="showCostTitle(until10Cost)">Cost: </span>
+        {{ format(until10Cost) }}
+      </template>
+    </PrimaryButton>
   </div>
 </template>
