+++ src/components/tabs/infinity-dimensions/ClassicInfinityDimensionsTab.vue
@@ -1,12 +1,12 @@
 <script>
 import PrimaryButton from "@/components/PrimaryButton";
-import OldInfinityDimensionRow from "./OldInfinityDimensionRow";
+import InfinityDimensionRow from "./ClassicInfinityDimensionRow";
 
 export default {
-  name: "OldInfinityDimensionsTab",
+  name: "ClassicInfinityDimensionsTab",
   components: {
     PrimaryButton,
-    OldInfinityDimensionRow
+    InfinityDimensionRow
   },
   data() {
     return {
@@ -138,7 +138,7 @@ export default {
     </div>
     <div>You are getting {{ format(powerPerSecond, 2, 0) }} {{ incomeType }} per second.</div>
     <div class="l-dimensions-container">
-      <OldInfinityDimensionRow
+      <InfinityDimensionRow
         v-for="tier in 8"
         :key="tier"
         :tier="tier"
