+++ src/components/tabs/time-dimensions/ModernTimeDimensionRow.vue
@@ -3,7 +3,7 @@ import PrimaryButton from "@/components/PrimaryButton";
 import PrimaryToggleButton from "@/components/PrimaryToggleButton";
 
 export default {
-  name: "OldTimeDimensionRow",
+  name: "ModernTimeDimensionRow",
   components: {
     PrimaryButton,
     PrimaryToggleButton
@@ -32,8 +32,13 @@ export default {
       requirementReached: false,
       realityUnlocked: false,
       showTTCost: false,
-      ttCost: 0,
+      ttCost: 0
     };
+  },
+  watch: {
+    isAutobuyerOn(newValue) {
+      Autobuyer.timeDimension(this.tier).isActive = newValue;
+    }
   },
   computed: {
     shiftDown() {
@@ -44,11 +49,13 @@ export default {
     },
     rateOfChangeDisplay() {
       return this.tier < 8
-        ? ` (+${format(this.rateOfChange, 2, 2)}%/s}/s)`
+        ? ` (+${format(this.rateOfChange, 2, 2)}%/s)`
         : "";
     },
     buttonContents() {
-      if (this.showTTCost) return this.formattedTTCost;
+      if (this.showTTCost) {
+        return this.formattedTTCost;
+      }
       return this.formattedEPCost;
     },
     tooltipContents() {
@@ -64,11 +71,6 @@ export default {
     },
     formattedEPCost() {
       return this.isCapped ? "Capped" : `Cost: ${format(this.cost, 2)} EP`;
-    }
-  },
-  watch: {
-    isAutobuyerOn(newValue) {
-      Autobuyer.timeDimension(this.tier).isActive = newValue;
     }
   },
   methods: {
@@ -115,7 +117,7 @@ export default {
     :class="{ 'c-dim-row--not-reached': !isUnlocked && !requirementReached }"
   >
     <div class="c-dim-row__label c-dim-row__name">
-      {{ name }} Time Dimension {{ formatX(multiplier, 2, 1) }}
+      {{ name }} Time D <span class="c-time-dim-row__multiplier">{{ formatX(multiplier, 2, 1) }}</span>
     </div>
     <div class="c-dim-row__label c-dim-row__label--growable">
       {{ format(amount, 2, 0) }}
@@ -129,7 +131,7 @@ export default {
     <PrimaryButton
       v-tooltip="tooltipContents"
       :enabled="isAvailableForPurchase && !isCapped"
-      class="o-primary-btn--buy-td l-dim-row__button"
+      class="o-primary-btn--buy-td l-dim-row__button o-primary-btn o-primary-btn--new"
       @click="buyTimeDimension"
     >
       {{ buttonContents }}
